Move mutators to another file
